requirements.txt
pylibagent==0.2.1
speedtest-cli==2.1.3
